more cleanups
