+++ .github/workflows/track-pr.yml
@@ -9,7 +9,7 @@ jobs:
     env:
       GITHUB_PR: ${{ toJSON(github.event.pull_request) }}
       PR_ID: ${{ github.event.pull_request.node_id }}
-      GH_TOKEN: ${{ secrets.GH_TOKEN }}
+      GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-go@v2
@@ -18,5 +18,4 @@ jobs:
 
       # Add PR to project board.
       - name: "Track PR"
-        if: ${{ success() }}
         run: go run testutil/pr/trackpr/main.go
